workspace: refactor libs into new structure
