+++ libs/ci-config/.babelrc
