+++ libs/ci-config/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@cypress-dx/ci-config",
+  "version": "0.0.1"
+}
