+++ libs/ci-config/src/lib/config.ts
