+++ libs/ci-generator/.babelrc
