+++ libs/ci-generator/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@cypress-dx/ci-generator",
+  "version": "0.0.1"
+}
