--- libs/ci/circle-ci/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@cypress-dx/circle-ci",
-  "version": "0.0.1"
-}
