--- libs/ci/circle-ci/src/index.ts
@@ -1,1 +1,0 @@
-export { generateCircleCI } from './lib/ci-circle-ci'
