--- libs/ci/config/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@cypress-ci/ci-config",
-  "version": "0.0.1"
-}
