--- libs/ci/generate/README.md
@@ -1,7 +1,0 @@
-# ci-generate
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test ci-generate` to execute the unit tests via [Jest](https://jestjs.io).
