--- libs/ci/generate/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@cypress-dx/ci-generate",
-  "version": "0.0.1"
-}
