--- libs/ci/github-actions/README.md
@@ -1,7 +1,0 @@
-# ci-github-actions
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test ci-github-actions` to execute the unit tests via [Jest](https://jestjs.io).
