--- libs/ci/github-actions/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@cypress-dx/github-actions",
-  "version": "0.0.1"
-}
