+++ libs/providers/circle-ci/.babelrc
