+++ libs/providers/circle-ci/.eslintrc.json
