+++ libs/providers/circle-ci/README.md
@@ -1,0 +1,7 @@
+# providers-circle-ci
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test providers-circle-ci` to execute the unit tests via [Jest](https://jestjs.io).
