+++ libs/providers/circle-ci/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  displayName: 'ci-github-actions',
+  displayName: 'providers-circle-ci',
   preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
@@ -10,5 +10,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/ci/github-actions',
+  coverageDirectory: '../../../coverage/libs/providers/circle-ci',
 };
