+++ libs/providers/circle-ci/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@cypress-dx/providers-circle-ci",
+  "version": "0.0.1"
+}
