+++ libs/providers/circle-ci/src/index.ts
@@ -1,0 +1,1 @@
+export { generateCircleCI } from './lib/ci-circle-ci';
