+++ libs/providers/github-actions/README.md
@@ -1,0 +1,7 @@
+# providers-github-actions
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test providers-github-actions` to execute the unit tests via [Jest](https://jestjs.io).
