+++ libs/providers/github-actions/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  displayName: 'ci-config',
+  displayName: 'providers-github-actions',
   preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
@@ -10,5 +10,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/ci/config',
+  coverageDirectory: '../../../coverage/libs/providers/github-actions',
 };
