+++ libs/providers/github-actions/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@cypress-dx/providers-github-actions",
+  "version": "0.0.1"
+}
