+++ libs/providers/github-actions/src/index.ts
@@ -1,0 +1,1 @@
+export { generateGitHubActions } from './lib/ci-github-actions';
